Add module name check logic (#384)
* Add module name check logic

* Better naming
